+++ src/components/App.js
@@ -19,14 +19,12 @@ import { HashRouter as Router, Route, Switch, Link } from "react-router-dom";
 import queryString from 'query-string';
 import HashChange from 'react-hashchange';
 
-import Github from './Github';
+import Github from '../data/Github';
 
-import Fail from './components/Fail';
-import Plan from './components/Plan';
-import Summary from './components/Summary';
-import Burndown from './components/Burndown';
-
-import './feature-dashboard.css';
+import Fail from './Fail';
+import Plan from './Plan';
+import Summary from './Summary';
+import Burndown from './Burndown';
 
 function getToken() {
     const GITHUB_TOKEN = 'github_token';
